chore: split gRPC microservice classes into multiple source files.
-Bridge Service now contained in BICBridgeGRPCService .cpp and .h files
-Device Service now contained in BICDeviceGRPCService .cpp and .h files
-BIC Device Info Repository structure now in BICDeviceInfoStruct.h
-BIC event listener class now in BICListener .cpp and . h files.
